requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
redis==5.0.7
neo4j==5.19.0
PyYAML==6.0.2